% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs.R
\name{drs_access}
\alias{drs_access}
\title{demonstrate acquisition of signed URL}
\usage{
drs_access(
  drs_uri = "drs://dg.4503:dg.4503/17141a26-90e5-4160-9c20-89202b369431"
)
}
\arguments{
\item{drs_uri}{character(1) DRS URI to be resolved}
}
\value{
character(1) signed URL
}
\description{
demonstrate acquisition of signed URL
}
\note{
the default value for drs_uri is a CCDG CRAM file reference; the associated
CRAI file has DRS URI `drs://dg.4503:dg.4503/1447260e-654b-4f9a-9161-c511cbdd0f95`
}
\examples{
drs_access()
}
