% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firecloud.R
\name{tables}
\alias{tables}
\alias{table}
\alias{table_delete_values}
\alias{table_upload}
\alias{table_gadget}
\title{Data table commands through firecloud}
\usage{
tables(namespace = tnu_workspace_namespace(), name = tnu_workspace_name())

table(
  table,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name()
)

table_delete_values(
  table,
  values,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name(),
  dry.run = TRUE
)

table_upload(
  tbl,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name()
)

table_gadget(
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name()
)
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace; see \code{?tnu_workspace}
for details and default value.}

\item{name}{character(1) AnVIL workspace name; see \code{?tnu_workspace} for
details and default value.}

\item{table}{character(1) table name.}

\item{values}{character(1) The '*_id' of the row that should be deleted from
the table.}

\item{tbl}{A tibble to uploaded as a data table to the workspace.}
}
\value{
\code{tables()} returns a tibble.

\code{table()} a tibble of data from the table.

\code{table_delete_values()} returns the name of the table that had the
row deleted from.

\code{table_gadget()} returns a tibble representing the selected table.
}
\description{
\code{tables()} lists the available tables in a workspace with the
corresponding row count and column names.

\code{table()} returns a table.

\code{table_delete_values()} deletes row from the table.

\code{table_upload()} uploads a table to the AnVIL workspace.

\code{table_gadget()} displays tables in the workspace and can be
returned as a tibble.
}
\examples{
if (tnu_workspace_ok()) {
    tables()
    table("test_table")
    mycars <- head(mtcars) |>
        dplyr::as_tibble(rownames = "model_id")
    table_upload(mycars)
    table_delete_values("model", "Mazda_RX4")
    table_gadget()
}

}
