% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_list}
\alias{table_list}
\alias{table_list_rows}
\alias{table_delete}
\alias{table_delete_row}
\alias{table_put_row}
\title{Data table commands through terra-notebook-utils}
\usage{
table_list(namespace = tnu_workspace_namespace(), name = tnu_workspace_name())

table_list_rows(
  table,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name()
)

table_delete(
  table,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name(),
  dry.run = TRUE
)

table_delete_row(
  table,
  row_id,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name(),
  dry.run = TRUE
)

table_put_row(
  table,
  item,
  namespace = tnu_workspace_namespace(),
  name = tnu_workspace_name()
)
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace; see \code{?tnu_workspace}
for details and default value.}

\item{name}{character(1) AnVIL workspace name; see \code{?tnu_workspace} for
details and default value.}

\item{table}{charater(1) AnVIL workspace table name.}

\item{dry.run}{A boolean to indicate if the deletion should happen. The
default is TRUE, meaning the deletion will not be executed.}

\item{row_id}{character(1) The row id(s) associated with the row(s) that
are to be deleted.}

\item{item}{A list of items to be added as a row in the specified table. If
adding multiple rows each row must be it's own list inside a list. See
example below.}
}
\value{
\code{table_list()} returns a character(1) list.

\code{table_list_rows()} returns a list of row information.

None

\code{table_delete_row()} returns the name of the table that had the
row(s) deleted from.

\code{table_put_row()} returns the id(s) of the row(s) added.
}
\description{
\code{table_list()} lists all available tables from a provided
workspace.

\code{table_list_rows()} lists all rows from a table.

\code{table_delete()} deletes a table from the AnVIL workspace.

\code{table_delete_row()} deletes a single (or multiple) rows from
the AnVIL workspace table.

\code{table_put_row()} adds a single (or multiple) rows to the AnVIL
workspace table.
}
\examples{
if (tnu_workspace_ok()) {
table_list()
}

if (tnu_workspace_ok()) {
table_list_rows("test_table")
}

if (tnu_workspace_ok()) {
table_delete("test_table")
}

if(tnu_workspace_ok()) {
table_delete_row("test_table", "d2098268-b3cc-4c22-b099-38c99586e196")

rws <- c("337af642-faa6-4eb6-a716-5645ccc84ec6",
    "98899664-1b73-4c14-9826-31760416ffb4")
table_delete_row("test_table", rws)
}

if (tnu_workspace_ok()) {
rw1 <- list(a = "z", b = "y", c = "x", d = 10)
rw2 <- list(a = "foo", b = "bar", c = "baz", d = 100)
table_put_row("test_table", rw1)
table_put_row("test_table", list(rw1, rw2))

rw3 <- list(list(a = "az", b = "by", c = "cx", d = 5),
    list(a = "du", b = "et", c = "fs", d = 50))
} 

}
