% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{tnu_drs_help}
\alias{tnu_drs_help}
\title{demonstrate use of basilisk to help on terra-notebook-utils DRS utilities}
\usage{
tnu_drs_help()
}
\description{
demonstrate use of basilisk to help on terra-notebook-utils DRS utilities
}
\examples{
tnu_drs_help()
}
