% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{tnu_help}
\alias{tnu_help}
\title{demonstrate use of basilisk to help on terra-notebook-utils}
\usage{
tnu_help()
}
\description{
demonstrate use of basilisk to help on terra-notebook-utils
}
\examples{
tnu_help()
}
