% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{tnu_top}
\alias{tnu_top}
\title{demonstrate use of basilisk to present features of terra-notebook-utils}
\usage{
tnu_top()
}
\description{
demonstrate use of basilisk to present features of terra-notebook-utils
}
\examples{
tnu_top()
}
