% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnu_utilities.R
\name{tnu_top}
\alias{tnu_top}
\alias{tnu_help}
\alias{tnu_drs_help}
\alias{tnu_table_help}
\title{Helpers to Discover terra-notebook-utils Functionality}
\usage{
tnu_top()

tnu_help()

tnu_drs_help()

tnu_table_help()
}
\description{
\code{tnu_top()} demonstrates use of basilisk to present
features of terra-notebook-utils

\code{tnu_help()} provides top-level help on
terra-notebook-utils python functions

\code{tnu_drs_help()} provides help on
terra-notebook-utils 'drs' module.

\code{tnu_table_help()} provides help on
terra-notebook-utils 'table' module.
}
\examples{
tnu_top()

tnu_help()

tnu_drs_help()

tnu_table_help()

}
