% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs.R
\name{drs_access}
\alias{drs_access}
\title{Access DRS through terra-notebook-utilities}
\usage{
drs_access(drs_uri)
}
\arguments{
\item{drs_uri}{character(1) DRS URI to be resolved}
}
\value{
character(1) signed URL
}
\description{
\code{drs_access()} acquires a signed URL from a DRS
address.
}
\note{
The URI used in the example is a CCDG CRAM file reference;
the associated CRAI file has DRS URI
\verb{drs://dg.4503:dg.4503/1447260e-654b-4f9a-9161-c511cbdd0f95}
}
\examples{
uri <- "drs://dg.4503:dg.4503/17141a26-90e5-4160-9c20-89202b369431"
if (tnu_workspace_ok()) {
    drs_access(uri)
}

}
